chore: migrate to TS useEditAsset
